Add files via upload
The BMI Calculator is a Python script designed to quickly and accurately calculate Body Mass Index (BMI) based on user-provided weight and height inputs. It offers a convenient way to assess an individual's body composition and categorizes the BMI result into standard weight categories, including Underweight, Normal weight, Overweight, and Obese. The script provides an intuitive command-line interface for easy interaction, making it accessible to users of all levels. With its simplicity and effectiveness, the BMI Calculator serves as a valuable tool for anyone seeking to monitor and maintain their health.